convert remaining cjs files to esm
