--- lib/index.cjs
@@ -1,37 +1,0 @@
-/**
- * Copyright 2013-2023 the original author or authors from the JHipster project.
- *
- * This file is part of the JHipster project, see https://www.jhipster.tech/
- * for more information.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-const { basename, join } = require('path');
-const { readFileSync } = require('fs');
-
-const getPackageRoot = () => {
-  const sourceRoot = join(__dirname, '..');
-  const sourceBasename = basename(sourceRoot);
-  return sourceBasename === 'generator-jhipster' ? sourceRoot : join(sourceRoot, '..');
-};
-
-const getPackageJson = () => {
-  return JSON.parse(readFileSync(join(getPackageRoot(), 'package.json')));
-};
-
-const packageJson = getPackageJson();
-module.exports = {
-  getPackageRoot,
-  getPackageJson,
-  packageJson,
-};
