+++ lib/support/needles.mjs
@@ -140,7 +140,7 @@ export const createNeedleCallback = ({ needle, contentToAdd, optional = false, i
 
   /**
    * @type {import('../../generators/base/types').EditFileCallback}
-   * @this {import('../../generators/base/index.cjs')}
+   * @this {import('../../generators/base/index.mjs').default}
    */
   return function (content, filePath) {
     contentToAdd = [].concat(contentToAdd).filter(eachContent => !checkContentIn(eachContent, content, ignoreWhitespaces));
