+++ test/support/index.mts
@@ -6,3 +6,4 @@ export * from './client-samples.mjs';
 export * from './server-samples.mjs';
 export * from './matrix-utils.mjs';
 export * from './matcher.mjs';
+export * from './temp-dir.mjs';
